Load all library files automatically in tests
Because it shouldn't be this easy to forget to load a new file. Some files in the libs dir were more a test subject so they were moved to src.
