+++ tests/Usages/AttributeUsagesTest.php
@@ -21,17 +21,18 @@ class AttributeUsagesTest extends RuleTestCase
 
 	protected function getRule(): Rule
 	{
-		$allowed = new Allowed(new Formatter(new Normalizer()), new Normalizer(), new AllowedPath(new FilePath(new FileHelper(__DIR__))));
+		$normalizer = new Normalizer();
+		$allowed = new Allowed(new Formatter($normalizer), $normalizer, new AllowedPath(new FilePath(new FileHelper(__DIR__))));
 		return new AttributeUsages(
 			new DisallowedAttributeRuleErrors($allowed, new Identifier()),
-			new DisallowedAttributeFactory($allowed, new Normalizer()),
+			new DisallowedAttributeFactory($allowed, $normalizer),
 			[
 				[
 					'attribute' => [
 						AttributeEntity::class,
 					],
 					'allowIn' => [
-						'../libs/ClassWithAttributesAllow.php',
+						'../src/disallowed-allow/ClassWithAttributesAllow.php',
 					],
 					'allowParamsAnywhereAnyValue' => [
 						[
@@ -48,15 +49,15 @@ class AttributeUsagesTest extends RuleTestCase
 	public function testRule(): void
 	{
 		// Based on the configuration above, in this file:
-		$this->analyse([__DIR__ . '/../libs/ClassWithAttributes.php'], [
+		$this->analyse([__DIR__ . '/../src/disallowed/ClassWithAttributes.php'], [
 			[
 				// expect this error message:
 				'Attribute Attributes\AttributeEntity is forbidden, because reasons',
 				// on this line:
-				10,
+				8,
 			],
 		]);
-		$this->analyse([__DIR__ . '/../libs/ClassWithAttributesAllow.php'], []);
+		$this->analyse([__DIR__ . '/../src/disallowed-allow/ClassWithAttributesAllow.php'], []);
 	}
 
 }
