+++ tests/bootstrap.php
@@ -2,13 +2,10 @@
 declare(strict_types = 1);
 
 require_once __DIR__ . '/../vendor/autoload.php';
-require_once __DIR__ . '/libs/Bar.php';
-require_once __DIR__ . '/libs/Blade.php';
-require_once __DIR__ . '/libs/Constructor.php';
-require_once __DIR__ . '/libs/Functions.php';
-require_once __DIR__ . '/libs/Inheritance.php';
-require_once __DIR__ . '/libs/Royale.php';
-require_once __DIR__ . '/libs/SomeInterface.php';
-require_once __DIR__ . '/libs/Option.php';
-require_once __DIR__ . '/libs/TestTrait.php';
-require_once __DIR__ . '/libs/Traits.php';
+
+$iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator(__DIR__ . '/libs', FilesystemIterator::SKIP_DOTS));
+foreach ($iterator as $fileInfo) {
+	if ($fileInfo->getExtension() === 'php') {
+		require_once $fileInfo->getPathname();
+	}
+}
